Dividir Game.cpp (audio, input y cámara) (#59)
* Refactorizar: extraer generación de audio a GameAudio.cpp

* Refactorizar: extraer gestión de input a GameInput.cpp

* Refactorizar: extraer lógica de cámara y FOV a GameCamera.cpp

* Pequeño error corregido

---------
