+++ src/core/Game.cpp
@@ -13,7 +13,7 @@ static inline unsigned now_seed() {
     return static_cast<unsigned>(time(nullptr));
 }
 
-static bool gQuitRequested = false;
+bool gQuitRequested = false;
 
 static Texture2D loadTex(const char *path) {
     const std::string full = assetPath(path);
@@ -28,21 +28,6 @@ static Texture2D loadTex(const char *path) {
     Texture2D t = LoadTextureFromImage(img);
     UnloadImage(img);
     return t;
-}
-
-static bool loadTextAsset(const char *relPath, std::string &outText) {
-    const std::string full = assetPath(relPath);
-    char *buf = LoadFileText(full.c_str());
-    if (buf) {
-        outText.assign(buf);
-        UnloadFileText(buf);
-        return true;
-    }
-    else {
-        outText = "No se encontro '" + full + "'.\n";
-        std::cerr << "[ASSETS] No se pudo cargar texto desde: " << full << "\n";
-        return false;
-    }
 }
 
 void ItemSprites::load() {
@@ -111,194 +96,6 @@ void ItemSprites::unload() {
     UnloadTexture(floor);
     
     loaded = false;
-}
-
-// Tipos de sonido para el generador
-enum SoundType { 
-    SND_HIT, SND_EXPLOSION, SND_PICKUP, SND_POWERUP, 
-    SND_HURT, SND_WIN, SND_LOOSE, SND_AMBIENT, SND_DASH
-};
-
-// Generador de ondas procedurales (Sintetizador)
-Sound Game::generateSound(int type) {
-    const int sampleRate = 44100;
-    
-    // CAMBIO: 60 segundos exactos.
-    const int durationFrames = (type == SND_AMBIENT) ? 44100 * 60 : 
-                               (type == SND_WIN || type == SND_LOOSE) ? 44100 * 4 : 
-                               44100 / 2;
-    
-    Wave wave;
-    wave.frameCount = durationFrames;
-    wave.sampleRate = sampleRate;
-    wave.sampleSize = 16;
-    wave.channels = 1;
-    wave.data = malloc(wave.frameCount * wave.sampleSize / 8);
-    
-    short *data = (short *)wave.data;
-    
-    for (int i = 0; i < wave.frameCount; i++) {
-        float t = (float)i / sampleRate; // Tiempo en segundos
-        float val = 0.0f;
-        
-        switch (type) {
-            case SND_HIT: // Ruido blanco corto + Onda cuadrada baja
-                if (t < 0.1f) val = ((float)GetRandomValue(-100, 100) / 100.0f) * (1.0f - t/0.1f);
-                else val = (sinf(2 * PI * 100 * t) > 0 ? 0.5f : -0.5f) * (1.0f - t);
-                break;
-                
-            case SND_EXPLOSION: // Ruido puro con decay largo
-                val = ((float)GetRandomValue(-100, 100) / 100.0f);
-                val *= (1.0f - (float)i / wave.frameCount); // Fade out lineal
-                break;
-                
-            case SND_PICKUP: // Onda cuadrada aguda (Coin)
-                // Frecuencia sube de 800 a 1200
-                {
-                    float freq = 800.0f + (t * 4000.0f); 
-                    val = (sinf(2 * PI * freq * t) > 0 ? 0.3f : -0.3f);
-                    val *= (1.0f - t*4.0f); // Muy corto
-                }
-                break;
-                
-            case SND_POWERUP: // Onda triangular subiendo (Powerup)
-                {
-                    float freq = 300.0f + (t * 1000.0f);
-                    val = asinf(sinf(2 * PI * freq * t)) * 0.5f; 
-                }
-                break;
-                
-            case SND_HURT: // Sierra descendente (Auch)
-                {
-                    float freq = 200.0f - (t * 400.0f);
-                    if (freq < 50) freq = 50;
-                    val = (float)(fmod(t * freq, 1.0) * 2.0 - 1.0) * 0.5f;
-                }
-                break;
-
-            case SND_AMBIENT: 
-                {
-                    // --- TEMA: "CYBER COMBAT LOOP" (60s Acción Pura) ---
-                    // Sin intro suave. Directo al grano.
-                    
-                    float bpm = 120.0f;
-                    float beatLen = 60.0f / bpm; // 0.5 segundos exactos
-                    
-                    // Variables de tiempo
-                    float measurePos = t / (beatLen * 4.0f); 
-                    int measure = (int)measurePos;           
-                    float beatPos = fmod(t, beatLen);        
-                    float subBeat = fmod(t, beatLen / 4.0f); 
-                    
-                    // --- 1. KICK (EL MOTOR) ---
-                    // Suena DESDE EL PRINCIPIO. Golpe seco y potente.
-                    float kick = 0.0f;
-                    if (beatPos < 0.2f) {
-                        // Frecuencia estable para que pegue fuerte
-                        float kFreq = 100.0f - (beatPos * 400.0f);
-                        if (kFreq < 40.0f) kFreq = 40.0f;
-                        kick = sinf(2 * PI * kFreq * t);
-                        kick *= expf(-15.0f * beatPos); 
-                        kick *= 0.9f; // Volumen alto
-                    }
-
-                    // --- 2. BASSLINE (LA ENERGÍA) ---
-                    // Bajo "Rolling" constante desde el segundo 0.
-                    float bassFreq = 55.0f; // La (A1)
-                    
-                    // Pequeña variación de nota cada 8 compases (16s) para no aburrir
-                    if ((measure / 8) % 2 == 1) bassFreq = 43.65f; // Fa (F1)
-
-                    // Onda de Sierra filtrada (agresiva pero controlada)
-                    float saw = (fmod(t * bassFreq, 1.0f) * 2.0f) - 1.0f;
-                    float bass = (saw * 0.4f) + (sinf(2 * PI * bassFreq * t) * 0.6f);
-                    
-                    // SIDECHAIN: El bajo se "agacha" cuando pega el Kick
-                    float sidechain = fmod(t, beatLen) / beatLen; 
-                    sidechain = powf(sidechain, 0.5f); // Curva rápida
-                    bass *= sidechain * 0.6f;
-
-                    // --- 3. HI-HATS (VELOCIDAD) ---
-                    // Ritmo de semicorcheas constante "tik-tik-tik-tik"
-                    float noise = ((float)GetRandomValue(-100, 100) / 100.0f);
-                    // Filtro High-Pass simulado (solo frecuencias altas)
-                    noise *= sinf(2 * PI * 6000.0f * t);
-                    
-                    float hat = noise * expf(-80.0f * subBeat) * 0.15f;
-                    // Acento en el "off-beat" (el "y" del 1 y 2...)
-                    if (fmod(t, beatLen) > beatLen/2.0f) hat *= 1.8f; 
-
-                    // --- 4. ARPEGIO (LA MELODÍA) ---
-                    // Entra y sale para dar variedad, pero el ritmo de fondo sigue.
-                    float arp = 0.0f;
-                    
-                    // La melodía suena en los segundos 0-15 y 30-45. 
-                    // Deja huecos de "solo ritmo" en 15-30 y 45-60.
-                    bool playArp = ((int)(t / 16.0f) % 2 == 0); 
-                    
-                    if (playArp) {
-                        int noteIdx = (int)(t * 8.0f); // 8 notas/segundo
-                        // Patrón simple y pegadizo
-                        float scale[] = { 1.0f, 1.0f, 1.5f, 1.2f, 2.0f, 1.5f, 1.2f, 1.5f };
-                        int idx = noteIdx % 8;
-                        float arpFreq = bassFreq * 4.0f * scale[idx];
-                        
-                        // Onda "Cristal"
-                        arp = sinf(2 * PI * arpFreq * t);
-                        // Delay
-                        arp += sinf(2 * PI * arpFreq * (t - 0.25f)) * 0.4f;
-                        
-                        float env = fmod(t, 0.125f) / 0.125f;
-                        arp *= (1.0f - env) * 0.15f; // Volumen sutil
-                    }
-
-                    // --- MEZCLA FINAL ---
-                    val = kick + bass + hat + arp;
-                    
-                    // Compresor/Limitador
-                    if (val > 0.9f) val = 0.9f;
-                    if (val < -0.9f) val = -0.9f;
-                    
-                    // Fade micro-rápido en los bordes del bucle (0.05s) para evitar "clicks"
-                    if (t < 0.05f) val *= (t / 0.05f);
-                    if (t > 59.95f) val *= (60.0f - t) / 0.05f;
-                }
-                break;
-            case SND_DASH: // Ruido rosa/aire rápido
-                {
-                    // Ruido aleatorio suave
-                    float noise = ((float)GetRandomValue(-100, 100) / 100.0f);
-                    // Volumen baja muy rápido (0.2 segundos)
-                    val = noise * 0.4f * (1.0f - t/0.2f); 
-                    if (val < 0) val = 0;
-                }
-                break;
-
-            case SND_WIN: // Acorde mayor arpegiado
-                {
-                    float note = 440.0f; // La
-                    if (t > 0.2) note = 554.37f; // Do#
-                    if (t > 0.4) note = 659.25f; // Mi
-                    if (t > 0.6) note = 880.0f;  // La (8va)
-                    val = sinf(2 * PI * note * t) * 0.5f;
-                    val *= (1.0f - t/2.0f);
-                }
-                break;
-
-            case SND_LOOSE: // Tono descendente triste
-                {
-                    float freq = 300.0f * (1.0f - t/2.0f);
-                    val = (sinf(2 * PI * freq * t) > 0 ? 0.4f : -0.4f);
-                }
-                break;
-        }
-        
-        data[i] = (short)(val * 32000.0f);
-    }
-    
-    Sound sound = LoadSoundFromWave(wave);
-    UnloadWave(wave);
-    return sound;
 }
 
 Game::Game(unsigned seed) : fixedSeed(seed) {
@@ -433,14 +230,6 @@ void Game::newLevel(int level) {
                                   rng, runCtx);
 }
 
-int Game::getFovRadius() const {
-    int r = fovTiles;
-    if (glassesTimer > 0.0f) {
-        r += glassesFovMod;
-    }
-    return std::clamp(r, 2, 30);
-}
-
 void Game::tryMove(int dx, int dy) {
     if (dx == 0 && dy == 0) return;
 
@@ -509,542 +298,6 @@ void Game::run() {
     CloseWindow();
 }
 
-void Game::processInput() {
-    if (state != GameState::MainMenu) {
-        bool escPressed = IsKeyPressed(KEY_ESCAPE);
-        const int gp0 = 0;
-        if (IsGamepadAvailable(gp0)) {
-            escPressed = escPressed ||
-                         IsGamepadButtonPressed(gp0, GAMEPAD_BUTTON_RIGHT_FACE_RIGHT) ||
-                         IsGamepadButtonPressed(gp0, GAMEPAD_BUTTON_MIDDLE_RIGHT) ||
-                         IsGamepadButtonPressed(gp0, GAMEPAD_BUTTON_MIDDLE);
-        }
-        if (escPressed) {
-            showHelp = false;
-            gAttack.swinging = false;
-            gAttack.lastTiles.clear();
-            state = GameState::MainMenu;
-            mainMenuSelection = 0;   
-            return;
-        }
-    }
-
-    auto restartRun = [&]() {
-        if (fixedSeed == 0) runSeed = nextRunSeed();
-        newRun();
-    };
-
-    if (IsKeyPressed(KEY_R)) {
-        restartRun();
-        return;
-    }
-
-    const int gpRestart = 0;
-    if (state != GameState::MainMenu &&
-        IsGamepadAvailable(gpRestart) &&
-        IsGamepadButtonPressed(gpRestart, GAMEPAD_BUTTON_RIGHT_FACE_UP)) {
-        restartRun();
-        return;
-    }
-
-    if (IsKeyPressed(KEY_T)) {
-        moveMode = (moveMode == MovementMode::StepByStep)
-                   ? MovementMode::RepeatCooldown
-                   : MovementMode::StepByStep;
-        moveCooldown = 0.0f;
-    }
-
-    if (IsKeyPressed(KEY_F2)) {
-        fogEnabled = !fogEnabled;
-        map.setFogEnabled(fogEnabled);
-        if (fogEnabled) map.computeVisibility(px, py, getFovRadius());
-    }
-
-    if (IsKeyPressed(KEY_LEFT_BRACKET)) {
-        fovTiles = std::max(2, fovTiles - 1);
-        if (map.fogEnabled()) map.computeVisibility(px, py, getFovRadius());
-    }
-    if (IsKeyPressed(KEY_RIGHT_BRACKET)) {
-        fovTiles = std::min(30, fovTiles + 1);
-        if (map.fogEnabled()) map.computeVisibility(px, py, getFovRadius());
-    }
-
-    const float dt = GetFrameTime();
-    if (state == GameState::MainMenu) {
-        handleMenuInput();
-        return;
-    }
-    if (state == GameState::Playing) {
-        handlePlayingInput(dt);
-        return;
-    }
-
-    player.update(dt, false);
-}
-
-void Game::handleMenuInput() {
-    const int menuGamepad = 0;
-
-    if (showHelp) {
-        // ... (Tu código de menú de ayuda, igual que antes) ...
-        int panelW = (int)std::round(screenW * 0.86f);
-        int panelH = (int)std::round(screenH * 0.76f);
-        if (panelW > 1500) panelW = 1500;
-        if (panelH > 900)  panelH = 900;
-        int pxl = (screenW - panelW) / 2;
-        int pyl = (screenH - panelH) / 2;
-
-        float wheel = GetMouseWheelMove();
-        if (wheel != 0.0f) {
-            helpScroll -= (int)(wheel * 40);
-            if (helpScroll < 0) helpScroll = 0;
-        }
-
-        if (IsGamepadAvailable(menuGamepad)) {
-            float ay = GetGamepadAxisMovement(menuGamepad, GAMEPAD_AXIS_LEFT_Y);
-            const float dead = 0.25f;
-            if (std::fabs(ay) > dead) {
-                const float speed = 400.0f; 
-                helpScroll += (int)(ay * speed * GetFrameTime());
-                if (helpScroll < 0) helpScroll = 0;
-            }
-        }
-
-        int margin    = 24;
-        int titleSize = (int)std::round(panelH * 0.06f);
-        int top       = pyl + margin + titleSize + 16;
-        int backFs    = std::max(16, (int)std::round(panelH * 0.048f));
-        int footerH   = backFs + 16;
-        int viewportH = panelH - (top - pyl) - margin - footerH;
-        int fontSize  = (int)std::round(screenH * 0.022f);
-        if (fontSize < 16) fontSize = 16;
-        int lineH = (int)std::round(fontSize * 1.25f);
-        int lines = 1;
-        for (char c : helpText) if (c == '\n') ++lines;
-        int maxScroll = std::max(0, lines * lineH - viewportH);
-        if (helpScroll > maxScroll) helpScroll = maxScroll;
-
-        const char *backTxt = "VOLVER";
-        int tw = MeasureText(backTxt, backFs);
-        int tx = pxl + panelW - tw - 16;
-        int ty = pyl + panelH - backFs - 12;
-        Rectangle backHit = { (float)(tx - 6), (float)(ty - 4),
-                              (float)(tw + 12), (float)(backFs + 8) };
-
-        bool clickBack = IsMouseButtonPressed(MOUSE_LEFT_BUTTON) &&
-                         CheckCollisionPointRec(GetMousePosition(), backHit);
-        bool escBack   = IsKeyPressed(KEY_ESCAPE);
-        bool gpBack    = false;
-
-        if (IsGamepadAvailable(menuGamepad)) {
-            gpBack = IsGamepadButtonPressed(menuGamepad, GAMEPAD_BUTTON_RIGHT_FACE_RIGHT) || 
-                     IsGamepadButtonPressed(menuGamepad, GAMEPAD_BUTTON_LEFT_FACE_RIGHT)  || 
-                     IsGamepadButtonPressed(menuGamepad, GAMEPAD_BUTTON_RIGHT_FACE_DOWN);    
-        }
-
-        if (clickBack || escBack || gpBack) {
-            showHelp = false;
-        }
-        return; 
-    }
-
-    int bw = (int)std::round(screenW * 0.35f);
-    bw = std::clamp(bw, 320, 560);
-    int bh = (int)std::round(screenH * 0.12f);
-    bh = std::clamp(bh, 72, 120);
-    int startY = (int)std::round(screenH * 0.52f);
-    int gap    = (int)std::round(screenH * 0.04f);
-
-    Rectangle playBtn = { (float)((screenW - bw) / 2), (float)startY,
-                          (float)bw, (float)bh };
-    Rectangle readBtn = { (float)((screenW - bw) / 2),
-                          (float)(startY + bh + gap), (float)bw, (float)bh };
-    Rectangle quitBtn = { (float)((screenW - bw) / 2),
-                          (float)(startY + (bh + gap) * 2), (float)bw, (float)bh };
-
-    if (IsMouseButtonPressed(MOUSE_LEFT_BUTTON)) {
-        Vector2 mp = GetMousePosition();
-        if (CheckCollisionPointRec(mp, playBtn)) {
-            newRun();
-            return;
-        }
-        if (CheckCollisionPointRec(mp, readBtn)) {
-            if (helpText.empty()) loadTextAsset("assets/docs/objetos.txt", helpText);
-            showHelp  = true;
-            helpScroll = 0;
-            return;
-        }
-        if (CheckCollisionPointRec(mp, quitBtn)) {
-            gQuitRequested = true;
-            return;
-        }
-    }
-
-    auto activateSelection = [&]() {
-        if (mainMenuSelection == 0) {
-            newRun();
-        }
-        else if (mainMenuSelection == 1) {
-            if (helpText.empty()) loadTextAsset("assets/docs/objetos.txt", helpText);
-            showHelp  = true;
-            helpScroll = 0;
-        }
-        else { 
-            gQuitRequested = true;
-        }
-    };
-
-    if (IsKeyPressed(KEY_UP) || IsKeyPressed(KEY_W)) {
-        mainMenuSelection = (mainMenuSelection + 3 - 1) % 3;
-    }
-    if (IsKeyPressed(KEY_DOWN) || IsKeyPressed(KEY_S)) {
-        mainMenuSelection = (mainMenuSelection + 1) % 3;
-    }
-    if (IsKeyPressed(KEY_ENTER) || IsKeyPressed(KEY_SPACE)) {
-        activateSelection();
-        return;
-    }
-
-    if (IsGamepadAvailable(menuGamepad)) {
-        if (IsGamepadButtonPressed(menuGamepad, GAMEPAD_BUTTON_LEFT_FACE_UP)) {
-            mainMenuSelection = (mainMenuSelection + 3 - 1) % 3;
-        }
-        if (IsGamepadButtonPressed(menuGamepad, GAMEPAD_BUTTON_LEFT_FACE_DOWN)) {
-            mainMenuSelection = (mainMenuSelection + 1) % 3;
-        }
-
-        static bool stickNeutral = true;
-        float ay = GetGamepadAxisMovement(menuGamepad, GAMEPAD_AXIS_LEFT_Y);
-        const float dead = 0.35f;
-
-        if (std::fabs(ay) < dead) {
-            stickNeutral = true;
-        }
-        else if (stickNeutral) {
-            if (ay < 0.0f) mainMenuSelection = (mainMenuSelection + 3 - 1) % 3;
-            else mainMenuSelection = (mainMenuSelection + 1) % 3;
-            stickNeutral = false;
-        }
-
-        if (IsGamepadButtonPressed(menuGamepad, GAMEPAD_BUTTON_RIGHT_FACE_DOWN)) {
-            activateSelection();
-            return;
-        }
-        if (IsGamepadButtonPressed(menuGamepad, GAMEPAD_BUTTON_RIGHT_FACE_UP)) {
-            if (helpText.empty()) loadTextAsset("assets/docs/objetos.txt", helpText);
-            showHelp  = true;
-            helpScroll = 0;
-            return;
-        }
-    }
-}
-
-// ----------------------------------------------------------------------------------
-// INPUT JUEGO (COMBATE Y MOVIMIENTO)
-// ----------------------------------------------------------------------------------
-void Game::handlePlayingInput(float dt) {
-    // --------------------------------------------------------
-    // 1. DETECCIÓN DE DISPOSITIVO
-    // --------------------------------------------------------
-    if (IsKeyDown(KEY_W) || IsKeyDown(KEY_A) || IsKeyDown(KEY_S) || IsKeyDown(KEY_D) ||
-        IsKeyDown(KEY_UP) || IsKeyDown(KEY_DOWN) || IsKeyDown(KEY_LEFT) || IsKeyDown(KEY_RIGHT) ||
-        IsKeyDown(KEY_E) || IsKeyDown(KEY_I) || IsKeyDown(KEY_O) || 
-        IsKeyDown(KEY_SPACE) || IsKeyDown(KEY_ONE) || IsKeyDown(KEY_TWO) || 
-        IsKeyDown(KEY_LEFT_SHIFT)) { // Añadido Shift
-        lastInput = InputDevice::Keyboard;
-    }
-
-    const int gpId = 0;
-    if (IsGamepadAvailable(gpId)) {
-        bool gpActive = false;
-        // Check botones
-        for (int b = GAMEPAD_BUTTON_UNKNOWN + 1; b <= GAMEPAD_BUTTON_RIGHT_THUMB; b++) {
-            if (IsGamepadButtonDown(gpId, b)) { gpActive = true; break; }
-        }
-        // Check ejes
-        if (!gpActive) {
-            float lx = GetGamepadAxisMovement(gpId, GAMEPAD_AXIS_LEFT_X);
-            float ly = GetGamepadAxisMovement(gpId, GAMEPAD_AXIS_LEFT_Y);
-            if (std::abs(lx) > 0.25f || std::abs(ly) > 0.25f) gpActive = true;
-        }
-        // Check gatillos
-        if (!gpActive) {
-             float rt = GetGamepadAxisMovement(gpId, GAMEPAD_AXIS_RIGHT_TRIGGER);
-             float lt = GetGamepadAxisMovement(gpId, GAMEPAD_AXIS_LEFT_TRIGGER);
-             if (rt > 0.1f || lt > 0.1f) gpActive = true;
-        }
-
-        if (gpActive) lastInput = InputDevice::Gamepad;
-    }
-
-    // --------------------------------------------------------
-    // 2. LÓGICA DE JUEGO (Zoom, Interacción)
-    // --------------------------------------------------------
-    if (IsKeyDown(KEY_I)) cameraZoom += 1.0f * dt;
-    if (IsKeyDown(KEY_O)) cameraZoom -= 1.0f * dt;
-
-    float wheel = GetMouseWheelMove();
-    if (wheel != 0.0f) cameraZoom = expf(logf(cameraZoom) + wheel * 0.1f);
-
-    if (IsGamepadAvailable(gpId)) {
-        float rightY = GetGamepadAxisMovement(gpId, GAMEPAD_AXIS_RIGHT_Y);
-        const float zoomThr = 0.2f;
-        if (rightY < -zoomThr) cameraZoom += (-rightY) * dt;
-        else if (rightY > zoomThr) cameraZoom -= rightY * dt;
-        
-        if (IsGamepadButtonPressed(gpId, GAMEPAD_BUTTON_RIGHT_THUMB)) {
-            cameraZoom = 1.0f;
-            camera.zoom = cameraZoom;
-            camera.rotation = 0.0f;
-            clampCameraToMap();
-        }
-    }
-
-    cameraZoom = std::clamp(cameraZoom, 0.5f, 3.0f);
-    camera.zoom = cameraZoom;
-    clampCameraToMap();
-
-    if (IsKeyPressed(KEY_C)) {
-        cameraZoom = 1.0f;
-        camera.zoom = cameraZoom;
-        camera.rotation = 0.0f;
-        clampCameraToMap();
-    }
-
-    // Interacción (Pickup)
-    bool interact = IsKeyPressed(KEY_E);
-    if (IsGamepadAvailable(gpId)) {
-        if (IsGamepadButtonPressed(gpId, GAMEPAD_BUTTON_RIGHT_FACE_DOWN)) interact = true;
-    }
-    if (interact) tryManualPickup();
-
-    // --------------------------------------------------------
-    // 3. DASH (ESQUIVA) - NUEVO BLOQUE
-    // --------------------------------------------------------
-    bool dashPressed = IsKeyPressed(KEY_LEFT_SHIFT) || IsKeyPressed(KEY_RIGHT_SHIFT);
-    if (IsGamepadAvailable(gpId)) {
-        if (IsGamepadButtonPressed(gpId, GAMEPAD_BUTTON_LEFT_TRIGGER_1) || 
-            IsGamepadButtonPressed(gpId, GAMEPAD_BUTTON_LEFT_TRIGGER_2)) {
-            dashPressed = true;
-        }
-    }
-
-    if (dashPressed && !isDashing && dashCooldownTimer <= 0.0f) {
-        int ddx = 0, ddy = 0;
-        // Prioridad: Dirección pulsada ahora
-        if (IsKeyDown(KEY_W) || IsKeyDown(KEY_UP))    ddy = -1;
-        if (IsKeyDown(KEY_S) || IsKeyDown(KEY_DOWN))  ddy = 1;
-        if (IsKeyDown(KEY_A) || IsKeyDown(KEY_LEFT))  ddx = -1;
-        if (IsKeyDown(KEY_D) || IsKeyDown(KEY_RIGHT)) ddx = 1;
-        
-        // Si no, última dirección
-        if (ddx == 0 && ddy == 0) {
-            ddx = gAttack.lastDir.x;
-            ddy = gAttack.lastDir.y;
-        }
-
-        if (ddx != 0 || ddy != 0) {
-            int targetX = px;
-            int targetY = py;
-            
-            // Calcular destino (hasta DASH_DISTANCE)
-            for (int i = 1; i <= DASH_DISTANCE; i++) {
-                int nextX = px + ddx * i;
-                int nextY = py + ddy * i;
-                if (!map.isWalkable(nextX, nextY)) break; // Pared
-                
-                bool enemyBlock = false;
-                for(const auto& e : enemies) if(e.getX() == nextX && e.getY() == nextY) enemyBlock = true;
-                if (enemyBlock) break; // Enemigo
-
-                targetX = nextX;
-                targetY = nextY;
-            }
-
-            if (targetX != px || targetY != py) {
-                // Configurar estado Dash
-                isDashing = true;
-                dashTimer = DASH_DURATION;
-                dashCooldownTimer = DASH_COOLDOWN;
-                
-                dashStartPos = { px * (float)tileSize, py * (float)tileSize };
-                dashEndPos = { targetX * (float)tileSize, targetY * (float)tileSize };
-                
-                px = targetX;
-                py = targetY;
-                player.setGridPos(px, py);
-                
-                PlaySound(sfxDash);
-                
-                onSuccessfulStep(0,0); 
-                return; 
-            }
-        }
-    }
-
-    // --------------------------------------------------------
-    // 4. MOVIMIENTO NORMAL
-    // --------------------------------------------------------
-    int dx = 0, dy = 0;
-    bool moved = false;
-
-    int gpDx = 0, gpDy = 0;
-    bool gpDpadPressed = false;
-    bool gpAnalogActive = false;
-
-    if (IsGamepadAvailable(gpId)) {
-        float axisX = GetGamepadAxisMovement(gpId, GAMEPAD_AXIS_LEFT_X);
-        float axisY = GetGamepadAxisMovement(gpId, GAMEPAD_AXIS_LEFT_Y);
-        const float thr = 0.5f;
-
-        if (axisX < -thr || axisX > thr || axisY < -thr || axisY > thr) {
-            gpAnalogActive = true;
-            if (axisY < -thr) gpDy = -1; else if (axisY > thr) gpDy = +1;
-            if (axisX < -thr) gpDx = -1; else if (axisX > thr) gpDx = +1;
-        }
-
-        if (IsGamepadButtonPressed(gpId, GAMEPAD_BUTTON_LEFT_FACE_UP)) { gpDx = 0; gpDy = -1; gpDpadPressed = true; }
-        if (IsGamepadButtonPressed(gpId, GAMEPAD_BUTTON_LEFT_FACE_DOWN)) { gpDx = 0; gpDy = +1; gpDpadPressed = true; }
-        if (IsGamepadButtonPressed(gpId, GAMEPAD_BUTTON_LEFT_FACE_LEFT)) { gpDx = -1; gpDy = 0; gpDpadPressed = true; }
-        if (IsGamepadButtonPressed(gpId, GAMEPAD_BUTTON_LEFT_FACE_RIGHT)) { gpDx = +1; gpDy = 0; gpDpadPressed = true; }
-
-        if (gpAnalogActive) moveMode = MovementMode::RepeatCooldown;
-        else if (gpDpadPressed) moveMode = MovementMode::StepByStep;
-    }
-
-    if (moveMode == MovementMode::StepByStep) {
-        if (IsKeyPressed(KEY_W) || IsKeyPressed(KEY_UP))    dy = -1;
-        if (IsKeyPressed(KEY_S) || IsKeyPressed(KEY_DOWN))  dy = +1;
-        if (IsKeyPressed(KEY_A) || IsKeyPressed(KEY_LEFT))  dx = -1;
-        if (IsKeyPressed(KEY_D) || IsKeyPressed(KEY_RIGHT)) dx = +1;
-
-        if (dx == 0 && dy == 0 && (gpDx != 0 || gpDy != 0)) { dx = gpDx; dy = gpDy; }
-        
-        if (dx != 0 || dy != 0) {
-            gAttack.lastDir = {dx, dy};
-            player.setDirectionFromDelta(dx, dy); 
-            int oldx = px, oldy = py;
-            tryMove(dx, dy);
-            moved = (px != oldx || py != oldy);
-            if (moved) onSuccessfulStep(dx, dy);
-        }
-        player.update(dt, moved);
-    }
-    else {
-        moveCooldown -= dt;
-        bool pressedNow = IsKeyPressed(KEY_W) || IsKeyPressed(KEY_S) ||
-                        IsKeyPressed(KEY_A) || IsKeyPressed(KEY_D) ||
-                        IsKeyPressed(KEY_UP) || IsKeyPressed(KEY_DOWN) ||
-                        IsKeyPressed(KEY_LEFT) || IsKeyPressed(KEY_RIGHT) ||
-                        gpDpadPressed;
-
-        if (IsKeyDown(KEY_W) || IsKeyDown(KEY_UP))    dy = -1;
-        if (IsKeyDown(KEY_S) || IsKeyDown(KEY_DOWN))  dy = +1;
-        if (IsKeyDown(KEY_A) || IsKeyDown(KEY_LEFT))  dx = -1;
-        if (IsKeyDown(KEY_D) || IsKeyDown(KEY_RIGHT)) dx = +1;
-
-        if (dx == 0 && gpDx != 0) dx = gpDx;
-        if (dy == 0 && gpDy != 0) dy = gpDy;
-
-        if (dx != 0 || dy != 0) {
-            gAttack.lastDir = {dx, dy};
-            player.setDirectionFromDelta(dx, dy);
-        }
-
-        if (pressedNow) {
-            int oldx = px, oldy = py;
-            tryMove(dx, dy);
-            moved = (px != oldx || py != oldy);
-            if (moved) onSuccessfulStep(dx, dy);
-            moveCooldown = MOVE_INTERVAL;
-        }
-        else if ((dx != 0 || dy != 0) && moveCooldown <= 0.0f) {
-            int oldx = px, oldy = py;
-            tryMove(dx, dy);
-            moved = (px != oldx || py != oldy);
-            if (moved) onSuccessfulStep(dx, dy);
-            moveCooldown = MOVE_INTERVAL;
-        }
-        player.update(dt, moved);
-    }
-
-    // --------------------------------------------------------
-    // 5. SISTEMA DE COMBATE (INPUT)
-    // --------------------------------------------------------
-    gAttack.cdTimer = std::max(0.f, gAttack.cdTimer - dt);
-    plasmaCooldown  = std::max(0.f, plasmaCooldown - dt);
-
-    if (gAttack.swinging) {
-        gAttack.swingTimer = std::max(0.f, gAttack.swingTimer - dt);
-        if (gAttack.swingTimer <= 0.f) {
-            gAttack.swinging = false;
-            gAttack.lastTiles.clear();
-        }
-    }
-
-    // A) MANOS
-    bool attackHands = IsKeyPressed(KEY_SPACE);
-    if (IsGamepadAvailable(gpId) && IsGamepadButtonPressed(gpId, GAMEPAD_BUTTON_RIGHT_FACE_LEFT)) attackHands = true;
-    
-    if (attackHands && gAttack.cdTimer <= 0.f) performMeleeAttack();
-
-    // B) ESPADA
-    bool attackSword = IsKeyPressed(KEY_ONE); 
-    if (IsGamepadAvailable(gpId) && IsGamepadButtonPressed(gpId, GAMEPAD_BUTTON_RIGHT_TRIGGER_1)) attackSword = true;
-    
-    if (attackSword) {
-        if (swordTier > 0) {
-            if (gAttack.cdTimer <= 0.f) performSwordAttack();
-        } else {
-            // std::cout << "No tienes espada\n";
-        }
-    }
-
-    // C) PLASMA
-    bool attackPlasma = IsKeyPressed(KEY_TWO);
-    if (IsGamepadAvailable(gpId)) {
-        if (IsGamepadButtonPressed(gpId, GAMEPAD_BUTTON_RIGHT_TRIGGER_2)) attackPlasma = true;
-    }
-    if (attackPlasma) {
-        if (plasmaTier > 0) {
-            if (plasmaCooldown <= 0.f) performPlasmaAttack();
-        } else {
-            // std::cout << "No tienes plasma\n";
-        }
-    }
-
-    // Cheats
-    if (IsKeyPressed(KEY_H)) { hp = std::max(0, hp - 2); }      
-    if (IsKeyPressed(KEY_J)) { hp = std::min(hpMax, hp + 2); }  
-}
-
-void Game::clampCameraToMap() {
-    const float worldW = map.width() * (float)tileSize;
-    const float worldH = map.height() * (float)tileSize;
-    const float viewW = screenW / camera.zoom;
-    const float viewH = screenH / camera.zoom;
-    const float halfW = viewW * 0.5f;
-    const float halfH = viewH * 0.5f;
-
-    if (worldW <= viewW) camera.target.x = worldW * 0.5f;
-    else camera.target.x = std::clamp(camera.target.x, halfW, worldW - halfW);
-
-    if (worldH <= viewH) camera.target.y = worldH * 0.5f;
-    else camera.target.y = std::clamp(camera.target.y, halfH, worldH - halfH);
-}
-
-void Game::centerCameraOnPlayer() {
-    camera.target = { px * (float)tileSize + tileSize / 2.0f,
-                      py * (float)tileSize + tileSize / 2.0f };
-    clampCameraToMap();
-}
-
-void Game::recomputeFovIfNeeded() {
-    if (map.fogEnabled()) {
-        map.computeVisibility(px, py, getFovRadius());
-    }
-}
-
 void Game::onSuccessfulStep(int dx, int dy) {
     if (dx != 0 || dy != 0) {
         gAttack.lastDir = {dx, dy};           
@@ -1054,13 +307,6 @@ void Game::onSuccessfulStep(int dx, int dy) {
     recomputeFovIfNeeded();                   
     centerCameraOnPlayer();                   
     updateEnemiesAfterPlayerMove(true);       
-}
-
-int Game::defaultFovFromViewport() const {
-    int tilesX = screenW / tileSize;
-    int tilesY = screenH / tileSize;
-    int r = static_cast<int>(std::floor(std::min(tilesX, tilesY) * 0.15f));
-    return std::clamp(r, 3, 20);
 }
 
 void Game::update() {
